files/requirements.txt
playwright==1.44.0
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.3
python-dotenv==1.0.1
sqlalchemy==2.0.30